ThriftServer lifetime management
Summary:
ThriftServers created for DuplexChannel seem to suffer from a shutdown problem. This is due to the fact that the ThriftServer cannot just stop the threads that are doing its execution, as such, any in-flight requests end up holding the Cpp2Connection alive (shared_ptr), that Cpp2Connection keeps the Cpp2Worker alive (shared_ptr). Unfortunately, the Cpp2Worker has a raw pointer back to the ThriftServer so it is crucial for it to not outlive the ThriftServer. Waiting for all in-flight requests to go away is also not an appealing solution for DuplexChannel usage since the creation and desctruction of ThriftServers happen during regular run-time not just start up and shutdown.

In order to fix the issue in the least-intrusive way possible, I added 2 new methods to ThriftServer and 1 new method to Cpp2Worker, only to be used for DuplexChannel cases, called startDuplex and stopDuplex. These are preferred methods to use when dealing with a client's ThriftServer. Most importantly, the stopDuplex method will initiate stop to the Cpp2Worker it has (it only has a single one since each client has its own ThriftServer) and passes a shared_ptr to itself to it. The Cpp2Worker stores this shared_ptr to the ThriftServer, guaranteeing that even if others drop their shared_ptrs to the ThriftServer, it will nt go away rendering the raw pointer in the Cpp2Worker invalid. The ThriftServer drops its shared_ptr to the Cpp2Worker once it passes its own shared_ptr to it, thus reversing the life time. As such, once the stopDuplex call is completed, the ThriftServer will no longer keep the Cpp2Worker alive but instead the Cpp2Worker will be holding the ThriftServer alive.

If the ThriftServer was the only thing holding the Cpp2Worker alive (the Cpp2Connection was already destructed when we invoked drainAllConnections, because there were no in-flight requests keeping it alive), then the Cpp2Worker will be destructed and the extra shared_ptr into the ThriftServer will also go with it. If there were in-flight requests, they will let go off the Cpp2Connection only when they are completed, which will let go off the Cpp2Worker and transitively the ThriftServer itself. None of the changes to Thrift should have any impact in other code as I did not introduce anything into the common paths. The only measurable impact might be the one new field I added to Cpp2Worker to hold the shared_ptr to the ThriftServer during the stop operation.

Reviewed By: haijunz, alandau

Differential Revision: D4390658

fbshipit-source-id: 5b95ad207377dbd8be50e1b2595190bddb9ffa4d