Perf regression for py3 in THeader
Summary:
_ord was 2x slower in py3 than py2.
And we called _ord 4 times in readFrame, now we do the ord operation once.

Also found that readAll was slower for small readAll operations, which we do
a bunch in readFrame.  I replaced the bytearray with list append and a final
join and smaller reads are more performant and larger reads are unchanged.

Reviewed By: haijunz

Differential Revision: D4400683

fbshipit-source-id: 9481b16404afe5fa37858a814323931d9b6f6989