SendRecvRequestCallback, for independent send and recv members
Summary:
[Thrift] `SendRecvRequestCallback`, for independent `send` and `recv` members.

The base `RequestCallback` has three members, `requestSent`, `requestError`, and `replyReceived`. The sequence of invocations of these members for various success and failure cases can be a touch confusing in some cases. In particular, for cases that need one after-`send` invocation and one after-`recv` invocation, it is a bit of a state machine. Let Thrift offer a `SendRecvRequestCallback`, a type derived from `RequestCallback`, to solve this case.

Also offer `FunctionSendRecvRequestCallback`, which is constructed with two `Function<void(/*...*/)>` arguments for `send` and `recv`. In particular, this makes testing easier. But it can also be useful separately in ad-hoc cases.

Reviewed By: haijunz

Differential Revision: D4315747

fbshipit-source-id: 2e7b02774034f2269e565093f5c658ebd870e5d3