Enable modular builds
Summary:
This enables to builds thrift in a granular way. You can specify if you want to build only the compiler or the libraries. If nothing is specified, everything will be built.

cmake flags:
* compiler_only=ON
* lib_only=ON
default flag:
* build_all=ON

I also updated the contbuild to try first the build with the compiler only with just the external dependencies that it needs. Then a contbuild with everything. Eventually I will add a contbuild with just the libraries where you can specify the path of the compiler.

Reviewed By: yfeldblum

Differential Revision: D4420687

fbshipit-source-id: f3db9fb7d5c16bb1b9b35af63a554b1063730b17