Remove flaky python tests
Summary:
Both TestCppServer.py and TestClientServer.py were running tests that depended on the initialization and destruction of the thrift server.

Before starting the tests, a thread is spawn to initialize the server.

Most of the times, the server will initialize fine and all the tests will pass. However, if we get in the scenario where we test one of the variables that is initialized by starting the server, then we will get a value that has not been set yet. These flaky tests are only present for functions that override TEventHandler. Besides, most of these tests were established for non-TCppServer servers which had a different initialization process. For TCppServer, it is not needed to test these things anymore.

Since I deleted the declaration from the *.thrift file, I also had to remove all the other overwritten functions. Note: All of those were empty implementations, those four functions were only used in TestCppServer.py and TestClientServer.py

Reviewed By: yfeldblum

Differential Revision: D4426622

fbshipit-source-id: a01f515058b155b55a32a4c89f02f46369c38ea8