Compiler: Add python tests to cmake
Summary:
Add the python fixture tests along with a fix for the ungenerated code. This helped uncover a bug that had been present in the open source builds.

Cmake was linking `generate_lib` instead of ${GENERATE_LINKED} which forces linking. Therefore, the compiler was dropping all the generator. The contbuild showed that the build was passing, which is true, but the compiler was not able to generate any code.

With this new python test, we need to actually use the compiler to generate some code. Thus, forching the change to ${GENERATE_LINKED}, otherwise it shows a contbuild error.

Reviewed By: yfeldblum

Differential Revision: D4421359

fbshipit-source-id: 9088373d71d660dc4220f64610e71669c47eb876