Fix skipping the file-contents comparisons in the compiler fixtures
Summary:
[Thrift] Fix skipping the file-contents comparisons in the compiler fixtures.

The bug is that, in `compare_code`, we were iterating a generator //twice//.

First time:

```lang=py
self.assertEqual(sorted(gens), sorted(fixt))
```

Second time:
```
for gen in gens:
```

This works with a list, but does not work with generators because any given generator may only be iterated once. Solution: make a list.

Reviewed By: eduardo-elizondo

Differential Revision: D4441435

fbshipit-source-id: a160b6ac392f2267bd84a37a3de7f8d3cb08b392