Kill SharedWheelTimer
Summary:
This class had a bad bug, where it was looking up the WheelTimer to use by the address of the EventBase.  The problem is, the global structure had no cleanup, so if you deleted an EventBase, and created another at the same address, you could retrieve an HHWheelTimer with a dangling pointer to the dead EventBase.

Turns out, someone added a built-in HHWheelTimer to EventBase, so use that when someone creates a WheelTimerInstance from an EventBase.

Note: A larger change is coming that will completely remove HHWheelTimer from the HTTPSession/HTTPConnector API, and always use the one in the EventBase.  This is a smaller change to mitigate the bug.

Reviewed By: lnicco

Differential Revision: D4436038

fbshipit-source-id: c046938fe61ea7b696529a590c222e6c2667c61d