fix ambiguous overloaded method
Summary:
This fixes an error from Cython that said "ambiguous overloaded method". The
problem is that strings and binaries are both treated as strings in C++, so if
you have a file that includes containers of both (ie: Set<string> and
Set<binary>), it will try to generate two identical move wrappers. This causes
the error.

My solution is to create a separate collection of container types that treats
binary and string identically for the purpose of generating those move wrappers.

Reviewed By: fried

Differential Revision: D4440087

fbshipit-source-id: 1648c89595bbe1b70d886326f246ade117d52d95