Fix two more client issues
Summary:
Issue 1: If the user submitted an iterable to a client where the client was
expecting a container, it would just send garbage to the server. The correct
action is to convert that iterable to the correct type of C++ object.

Issue 2: An import was missing, which caused enums to not be processed
correctly. We had the correct cimport statement for local types, but not the
associated import statement, which permits access to python functions in the pyx
file.

Reviewed By: fried

Differential Revision: D4443016

fbshipit-source-id: 28edec2062d2597472caf3ccc2d91408d271bf67