Hack arrays support
Summary:
This diff adds support for Hack arrays (dict, vec, keyset) to the
thrift generator, as an alternative to collections. This is opt in, and should
be compatible with generated thrift that does still uses collections.

This diff adds support for the thrift generator, as well as updates the thrift
extension for HHVM to produce hack arrays.

Reviewed By: eduardo-elizondo

Differential Revision: D4402769

fbshipit-source-id: 655cfb7bed21742d93dc7152da4e85bd03eb5e3d