Support indirection over string in cpp2
Summary:
[Thrift] Support `indirection` over `string` in `cpp2`.

We already support `indirection` over numeric types, and the test suite validates the support for `indirection` over `i32`. And we currently support `indirection` over `string` when compiling with `cpp`. But we do not support it when compiling with `cpp2`.

The patches in this commit to `thrift/test/fatal_reflection_indirection_types.h` and `thrift/test/fatal_reflection_indirection.thrift` work when compiling with `cpp`. They fail to compile with `cpp2`, however, with this compiler error:

```
thrift/test/gen-cpp2/fatal_reflection_indirection_types.cpp:24:17: error: no viable overloaded '='
  phrase.phrase = CppHasAPhrase();
  ~~~~~~~~~~~~~ ^ ~~~~~~~~~~~~~~~
```

So we support to the `cpp2` compiler for this case and we add a unit-test to validate the support.

Note: The code being changed has nothing to do with reflection per se, but I am adding the tests to the test suite for reflection over indirection because it exists, as of recently anyway.

Reviewed By: eduardo-elizondo

Differential Revision: D4514309

fbshipit-source-id: 0fa4ada902aa1b78f7cee338f935c86d8dbb4849