Generate cpp where they are expected
Summary:
The cpp/.h files are now generated where they are expected, instead of
relying on a trick of the build process to make sure they are in the right place.

There will still need to be some trickery for _api.h resulting from
cython compilation.

I regenerated fixtures several times but there were no changes. So we need to
expand our fixtures to include a test with a namespace defined.

Reviewed By: buchuki

Differential Revision: D4457933

fbshipit-source-id: 3db235cae56210568e3e4a944b55254a959a321f