Fix invocation of pack to use a byte string as the input arg
Summary:
mobileconfig-generator-osx.par is being executed as part of the fbobjc messenger build and it's failing on devservers.
I believe this post describes a fix (due to use of __future__ unicode_literals), and seems reasonable:
https://www.facebook.com/groups/fbpython/permalink/631082610266443/

Found more info. This issue is in python's bug tracker, and explains how this works on newer python 2.7's.
https://bugs.python.org/issue19099

Reviewed By: mzlee

Differential Revision: D4516020

fbshipit-source-id: b04729e32d89218dcaca5c408bdef5472d4916ce