Don't attempt to import services if there aren't any
Summary:
In an earlier diff, fried removed empty services and types files if there is no
code to be generated in them. However, we were still attempting to import those
files in the generated code. This diff special cases those imports.

In addition, there are several fixes related to using enums with containers;
these were previously untested and not working.

Reviewed By: yfeldblum

Differential Revision: D4530004

fbshipit-source-id: 01684a811e2c746b38ee2870efe4843a26c1fa47