types robustness
Summary:
Introduced thrift.py3.types
'Exception' and 'Struct' from that module are the base class for all
thrift structs and exceptions.

I added __call__ to structs, to behave like namedtuple._replace.
We should discurage the use of Struct classes directly and point people
at the new {{struct:name}}_defaults which is what we use when you pass a
None for one of the __call__ args.  Maybe in the future we can move
structs to _{{struct:name}} then {{struct:name}} could be the defaults instance.
Since structs are immutable, if you use call and don't change anything you get
the same object back. (We could expand that to see if the thing you are setting
is the same value, in that case we could return the same object also)

Added __repr__ impls to all generated type classes.  These are very basic,
with no logic for limiting the repr string ala pretty print.

With this I made use of __isset to show None to python users for fields that
had no value, but did not also have a default defined in .thrift. This gives
out structs a more native feel.

I ran through all the container classes we build and made sure they can't segfault
python.  Mostly I did this by providing quick escapes when the container is
empty.

All classes now support weakrefs being made of them, and they all have
cached __hash__ functions.

Reviewed By: buchuki

Differential Revision: D4540019

fbshipit-source-id: 9800183e07f66402610c58341cefd98b14d80c48