Use insert instead of emplace for the set specialization of from_dynamic
Summary:
[Thrift] Use `insert` instead of `emplace` for the set specialization of `from_dynamic`.

Some custom set types do not have `emplace`, but they should generally have `insert`.

And the way `emplace` is used here is equivalent to `insert`, anyway, so `emplace` offers no benefit here relative to `insert`.

(Note: We have traits which must be specialized for the various custom collection types and this code is not using them. That is a bug, but that would have to be handled separately.)

Reviewed By: wqfish

Differential Revision: D4584541

fbshipit-source-id: a11d033c5a9b19b5ce7d6999785331026a8c31b1