Add a threadmanager that uses a proper priority queue instead of managing multiple thread pools
Summary:
The current default thread manager, PriorityThreadManager, has a number of issues:

1) It plays with kernel priorities, which is often not a great idea
2) Tuning the high pri vs low pri thread pools is difficult, and results excess threads that can contend
3) GetStatus runs on a different thread pool than normal requests by default, meaning deadlocks aren't detected by getStatus
4) a combination of 2/3, getCounters and getStatus can saturate their tiny (by default) threadpool and fail spuriously. While not many resources should be spent on these calls, it's better to show up as inefficiency when they are rather than failure.

And some more, but hopefully this is enough to justify. This adds a priority queue. Note that this diff actually makes the default TM basically support priorities, but it doesn't expose an interface. There's not much efficiency impact as the priority queue is just a thin wrapper on MPMCQueue when only one priority is available. A future diff will switch this to default

Reviewed By: djwatson

Differential Revision: D4460670

fbshipit-source-id: dc973e4643fbe055353093863e319599f803d193