fix structs with no fields
Summary:
iter on a struct or exception with no fields will be empty.
bool on a struct or exception with no fields will be True.

Exceptions will use the hash method of their parent type, because in python
even exceptions of the same type with same arguments have different hash values.

Hash of structs with no fields will be the hash of the Class.

Reviewed By: buchuki

Differential Revision: D4622659

fbshipit-source-id: c0a70dbba590925802e5a0cef6610e18e3496994