Add getOptional() for associative containers
Summary:
We already have `getDefault()`, but it's not useful in cases where a
default doesn't make sense. The only interface other than `find()` +
compare against `end()` we have is a throwing `at()`, which also isn't
reasonable a lot of times.

This adds a `getOptional()` method that allows to write concice, readable
code, e.g.:

    auto it = frozen.find(key);
    if (it != frozen.end()) {
      foo(it->second);
    }

becomes:

    if (auto v = frozen.getOptional(key)) {
      foo(*v);
    }

Reviewed By: spacedentist

Differential Revision: D4659828

fbshipit-source-id: e564749303512b15bcfd1a27b6a0eb6399e225ee