Moved thrift dependencies that clash on windows
Summary:
Some of the thrift dependencies don't exist on Windows and importing
them at the top of a file causes that file and all files importing them to
instantly crash on Windows. Some of this dependencies are only used in one
method and there's no point in not being able to use a whole file because of
this so just moving the import to the local method.

Reviewed By: phyllipecesar

Differential Revision: D4666352

fbshipit-source-id: 70034bee11ad8893d0bd60a5ca84987045c482e5