Add hook for client identity extraction
Summary:
This change adds a function to the thrift server that allows for
secure client identities to be stored in the connection context. The hook can
be set with setClientIdentityHook. When the connection is created, the hook
will be called with the peer cert and sasl server as arguments. The resulting
object is stored in the connection context.

Reviewed By: haijunz

Differential Revision: D4680057

fbshipit-source-id: d8fd2d698870cbe850b506734d294e9359695382