Better control of timing in the deserialization bench
Summary:
[Thrift] Better control of timing in the deserialization bench.

By using the suspender method that takes a lambda expression to be included in the timing.

Specifically, the prior code forgot to `rehire()` the suspender before dtors ran, and dtors in this benchmark do nontrivial work and affect the benchmark results. By constructing a suspender at the topmost scope and first, we implicitly exclude all dtors from timings. And by using the `dismissing()` method taking a lambda expression, it is clear that only the specified code is included in the timings, and we exclude dtors instead of accidentally include them.

Reviewed By: andrewjcg

Differential Revision: D4777627

fbshipit-source-id: 46128edf34b03d5830ad6a9eccf8dc9fce95442a