Change SimpleJSON python code to support subclassing
Summary: Make TSimpleJSON use class attributes for the contexts, to make it easier to subclass. Essentially, the Context classes are now assigned to attributes of the ProtocolBase, and all cases where the Context was previously a reference to a class in the same module have become references to the class attributes. This enables the inheriting class to change them for Contexts that don't add whitespace.

Reviewed By: yfeldblum

Differential Revision: D4745710

fbshipit-source-id: 3d780b43f07e34787cae388df63309a2b0dae925