Compiler: Remove container deserialization from tests
Summary:
Unit tests compare cpp2 to cpp2_mstch. There will be a change to the deserialization code so we need to remove this temporarily to make all unit tests pass.

The other option was to fix the mustache templates to account for this change but it was going to add up to an already big diff.

Reviewed By: yfeldblum

Differential Revision: D4826619

fbshipit-source-id: cff7562ddb4ba569a7e160139d963a6040680c19