Add protocol_methods to resolve container de/serialization
Summary:
This code moves code from thrift/lib/fatal/serializer.h to thrift/lib/serializer.h to be able to reuse without reflection.

This also adds new code (marked inline) to fix corner cases:
* enum classes
* fallback to Cpp2Ops for type_class::structure

This is a modification to the fatal implementation of serializer and does not use reflection capabilities so it's unable to read Structs. Instead, this is used to read any type of container and primitive types

Reviewed By: yfeldblum

Differential Revision: D4826672

fbshipit-source-id: f15746e9351bf0e643f320ed993afa9bf4707ea8