thrift-py3 client inheritence context manager warning fix
Summary:
We have this simple test to determine if the context manager was used
properly, well for inheritence clients the parent classes still had
a valid pointer to the c++ wrapper, and their deallocs would print a
warning even though C++ would not abort, because we cleaned that up correctly.

This fixes that warning from printing.

Reviewed By: ambv

Differential Revision: D4871246

fbshipit-source-id: 603bdb4d61bbc3784fd889fc7f00505f1c91d181