fix reentrancy bug in HeaderClientChannel
Summary:
We had a panicking customer -
https://www.facebook.com/groups/fbthrift/permalink/1698559613504267/ - seeing
crashes. We tried to help and we found a problem in our code.

Consider this callstack - P57261681$9,12,60,73,74
What is going on there:
There were (at least) two requests were placed into a HeaderClientChannel.
We are in a state where latter request is still queued up for send.
At the same the earlier request has encountered an error (EOF) and we need to
error out all pending requests.
This is this line in the callstack - P57261681$163

Looking at the code in HeaderClientChannel::messageReceiveErrorWrapped (link
diffusion/FBS/browse/master/fbcode/thrift/lib/cpp2/async/HeaderClientChannel.cpp;0da9ac1b4b7ee6bc3daaf44aa3fe73f6a8051b68$588-597), we are swapping out all pending recv callbacks from channel's member
variable onto a stack variable, and then invoke requestError on each one of the
callbacks.

The very first callback comes back and decides to close the DuplexChannel that
is corresponding to the call - P57261681$153

As part of the cleanup, we need to process the send error for the request that
is still pending to be sent out - P57261681$113-114

The pending send needs to remove the receive callback from the
HeaderClientChannel and it tries to find this callback by seqId. However, since
the callbacks have been already swapped out onto the stack, erasing fails with
a CHECK.

The fix here is to avoid swapping out ALL pending recv callbacks from member variable into a stack variable. Instead we remove and invoke them one-by-one. Since invoking the callback may result in channel closing and invoking the rest of the callbacks (while removing those from the list as well), we continue iterating on the list until it becomes empty.

Differential Revision: D4812845

fbshipit-source-id: 740d31474337741c92f7ee90e1e647ff745af33d