Compiler: Cache already parsed files - O(2^n) -> O(n)
Summary:
This changes the internal representation of the thrift AST data structure and adds caching for every file already parsed.

t_program changes:
Previously, for every include, we created a new t_program. With this new change, we added a map that holds all the already created t_program using the program path as a key. If a program is required and it has already been created, then we will just access the map and return a pointer to the already created program.

t_scope changes:
Previously, for every program we had a t_scope object that handled all the scopes that were included in child elements. With this new change, we now created a t_scope cache object that will hold every type, constant, and service with the correct namespace.

t_typedef changes:
Previously, t_typedef would look into the program scope to determine it's t_type. However, we now use global scope. This also modifies the constructor of t_typedef to accept the scope to correctly resolve its t_type.

Finally, we modified the way that we added elements to the scope. Previously, if the element was defined within the current program, then it would not get a unique name depending on its file name. This caused some name conflicts with different thrift files that had the same name for structs. Since we now have a global scope, we now uniquely identify each of those by their file name.

This reduced the parsing and generation that had to be done from O(2^n) to O(n). Speedups are almost only relevant in thrift files with many includes. However, some benchmarking showed that this reduced the codegen time of a thrift file from 225 seconds to 1.5 seconds. A 150x speedup.

Differential Revision: D4889029

fbshipit-source-id: 211379dd27eb10eddec6a0218e1db961a1800bc6