Expose line numbers information in thrift compiler
Summary:
Motivation: this will be useful in tools that does refactoring of thrift files
as we need to understand how AST corresponds to actual file.

In addition we export this values in json_generator as a way to confirm it works
Currently we only dump line numbers information for top-level definitions, but we might annotate fields as well later on.

Reviewed By: yfeldblum

Differential Revision: D4892937

fbshipit-source-id: f9112a40442603a448669b74ad2e1154dceb48fa