Reflection headers should include dependencies in a stable order
Summary:
[Thrift] Reflection headers should include dependencies in a stable order.

Instead of depending on the internal details of `sorted`, which may change between versions of Python, let us choose precisely how we sort, which will not change between versions of Python.

The problem this solves is that we may generate the test fixtures with one version of Python but check them with `compiler_test.py` with another version of Ptyhon, causing the unit-test to fail.

Reviewed By: andrewjcg

Differential Revision: D4990234

fbshipit-source-id: 54230e5373e68a4dc7383feb470094343dfff67b