Views aren't pointer-like, but iterators are
Summary:
I was somewhat indecisive in whether Views should be pointer-like or not. It turns out that making them values and not iterator-like constructs works much better. I've removed `ViewBase::operator->()`, which just returned `this`, and fixed all the compilation errors which resulted.
Depends on D4922789

(Note: this ignores all push blocking failures!)

Reviewed By: philippv

Differential Revision: D4931833

fbshipit-source-id: 13ccc5fdeca20d3608b3bf1a3dc185c722940333