thrift-py3 partial template re-organize
Summary:
Trying to organize the template files so they are easier to reason about.
The current scheme is templates that genetate resulting files are left in the
top level directory, everything else is organized in a directory for where
they are used.

I also renamed all the top level templates so they match the name of the
resulting compiled file to lower confusion.

Reviewed By: yfeldblum

Differential Revision: D4990995

fbshipit-source-id: a40b35c6933babcb95c2ee3a867375308fac784a