Fix use of pointer-to-temporary for in key comparison
Summary:
`KeyExtractor::getKey`, when used with `VectorAsMap`, would return a reference
to a temporary `pair<const X, Y>` that gets constructed from the `pair<X, Y>`
being passed in. A pointer to this data was stored and used for subsequent
distinctness checks, which would then fail.

This adds an overload to support non-const keys.

(Note: this ignores all push blocking failures!)

Differential Revision: D5002746

fbshipit-source-id: cb5cf671362c4906e5df6f9c46f8758b1d6c090a