Remove the template from the cpp2 indirection helpers
Summary:
[Thrift] Remove the template from the cpp2 indirection helpers.

They do not strictly need to be templates and it would be easier to use them in other places were they not. Instead, we move the template, which is needed for detection of types with indirections for serialization, into the serialization helpers.

And, to help with future reuse, we move the definitions into the header. They are small and rare.

Reviewed By: eduardo-elizondo

Differential Revision: D5017485

fbshipit-source-id: 04381c0a865feaabf39ffe41fd3d7240ca9c5ed7