Compute the default value in generate_php_struct_shape_spec using the same method used elsewhere
Summary: Prior to this diff, `generate_php_struct_shape_methods` computed the default value for the shape incorrectly in the case where the field had a default value. If the field has a default value, the shape was computed as being nullable, whereas in reality, it is not nullable. This diff fixes the bug by copying and pasting the logic to compute the default value that is used elsewhere within the file (e.g., in `generate_php_struct_shape_methods`) to be the same logic used in `generate_php_struct_shape_methods`.

Reviewed By: eduardo-elizondo

Differential Revision: D5088960

fbshipit-source-id: 669478c5a67d272fc9c1c1ef03b1771a1c5cfcca