Remove unnecessary code in the to_dynamic tests
Summary:
[Thrift] Remove unnecessary code in the `to_dynamic` tests.

In this case, the param `json` is unused, so remove it. And the two halves of the function are marked as testing both varieties of `format_adherence`, but `format_adherence` is only used with `from_dynamic` and is not used with `to_dynamic`, so the two halves of the function just test the same thing, so deduplicate.

Reviewed By: andrewjcg

Differential Revision: D5090786

fbshipit-source-id: 2b1bcde0f818cd31422e8edfb66e7405c3b3eb71