Compiler: Modify AST to add enum context to t_const_value
Summary:
This modifies t_enum_value to have a pointer to the enum parent. In this way, we will be able to always resolve the full enum name given just the t_enum_value node.

This also modifies t_const_value to add an is_enum bool and a t_enum_value pointer. When the t_const_value is being created, if it comes from an enum, then this information will be passed in. This will help us distinguish from a normal integer, to an integer that comes from an enum_value.

Reviewed By: yfeldblum

Differential Revision: D5123668

fbshipit-source-id: 3f1a50f0922fc25e04f3954290674c6f21acb2c7