Simplification of type-related partials
Summary: There was a fair amount of redundancy between different partial definitions - in particular, there were five different partials for declaring a Cython type equivalent of a C++ type, and they were all identical except that some of them were namespaced and some weren't. I added some code to check if you were in a types.pyx/pxd file or not (which changes the namespacing rule) and this enabled consolidation down to one template.

Reviewed By: fried

Differential Revision: D5209675

fbshipit-source-id: 704c5a2654265ed4a6df282274559372a024c3dc