pass_context for server handler methods
Summary:
Exposing a pass_context decorator that can be used
with async handler methods, which will cause the method to
always be called with a context parameter right after 'self'.

Right now we are exposing the Cpp2RequestContext, it only has
a single property to access the 'connection_context' which exposes
peer_address, peer_certificate, is_tls, peer_common_name properties.

In the future we will expost connection and request headers via this
interface.

Reviewed By: danshiovitz

Differential Revision: D5232810

fbshipit-source-id: 990a4269ba81a0e789e6ab68d60f40dedb938196