Introduce lazy_constants initialization for PHP Thrift
Summary: This functionality was introduced in the Hack generator in D2789963. When working on some fixes for the Hack/PHP compilers I noticed that some test fixtures were not valid code because they create new objects and therefore need lazy initialization. So I introduced this code to ensure that all the test fixtures for PHP would also be syntactically valid PHP.

Reviewed By: yfeldblum

Differential Revision: D5387076

fbshipit-source-id: 55b993b489ccba839de128b84638fd4beee5fd43