Create map for enum types
Summary:
This adds a two variable for all enum, a EnumToName and EnumToValue map.  This will allow you to iterate throug all the values / names of a enum.

Since this is similar to the Sting() and FromName() functions they have been updated to to use the maps instead of case statements.

For example:

    for val, name := range MyEnumToName {
       fmt.Printf("%d = %s\n", val, name)
    }

Reviewed By: fiorix

Differential Revision: D5380475

fbshipit-source-id: b46c3e3eedd99e7c775ad5eb90439046034501ef