Remove ClientReceiveState::exception
Summary:
[Thrift] Remove `ClientReceiveState::exception`.

Prefer instead, in the general case, to use `ClientReceiveState::exceptionWrapper()`, which returns `folly::exception_wrapper` instead of `std::exception_ptr`. Unlike `ClientReceiveState::exception()`, it can frequently avoid the internal throw-and-catch that needs to be done when dealing with a `std::exception_ptr`, which makes it generally more efficient.

The `std::exception_ptr` that used to be gotten from an invocation of `ClientReceiveState::exception()` can now, instead, be gotten by code that really needs it as `ClientReceiveState::exceptionWrapper().to_exception_ptr()`.

Reviewed By: GeniusYe

Differential Revision: D5431994

fbshipit-source-id: 849a9e54626a714e878c18b1569f381691c1d4be