mstch-cpp2: Fix typedef same as type
Summary: If we typedef with the same name as the type we would get a forward declare of the typedef name which would cause conflict during compilation. This fixes that. Ideally, we should not have typedefs doing this in thrift but some languages (i.e. swift) need this hack because of lack of support

Reviewed By: yfeldblum

Differential Revision: D5466114

fbshipit-source-id: 8d80f6035a4ccebdf35903e9dfce05fc00953451