Show error messages properly in arg parsing
Summary:
This just updates how we check whether an argument was provided for an option on the command line. It still requires the '--option' to be the last thing, and for no file to be provided, but at least in those cases, the user will get a warning.

The argparse logic before was strange, but would never be an actual problem because we only go to argc-1. However, it means that some useful error messages might never get displayed. This doesn't make it perfect by any means, but it does make the code at least reachable.

Reviewed By: yfeldblum

Differential Revision: D5441149

fbshipit-source-id: 9d6cfd52986400be8b18c904905cf78d2328ea7e