Add support for unions
Summary: Add support for unions. A thrift union Foo produces a class Foo as usual and also an enum FooType. Instances of unions have two members, one that holds the value and one that holds the type, which get loaded on object creation (this breaks our lazy-loading paradigm a little bit, but really, to do anything useful with a union you're going to need to load the (single) value).

Reviewed By: fried

Differential Revision: D5472196

fbshipit-source-id: cbeacf1066c0dafc0a76fc3568852cb5beec1d31