Move cpp.type from union to typedef
Summary:
cpp.type should only be used in a typedef or a field not a struct/union. Thus, cpp.type does not have an effect on this particular union - therefore it is misplaced.

When attempting to port to mstch-cpp2, it turned out that t_cpp_generator would check if the struct/union contained a cpp.type. This is not consistent, we should only expect a cpp.type from a typedef or a field. This moves the cpp.type to a typedef to be consistent.

Reviewed By: yfeldblum

Differential Revision: D5481401

fbshipit-source-id: f35446de26429dbb4769823ff6a4214087f1c8e6